Split push plugin up into multiple files
